fix(squashfs): compile against vfs, and give readdir its index back
Three type errors kept `cargo check -p squashfs` red on main. Nothing in
the kernel build depends on this crate, so `hosted-gate` was the only gate
that could see it at all. Fixed at the callers: the timestamp type's
fields are `sec`/`nsec` here, and `f_namelen` is a `u64`.

Fixing that exposed two dead methods, and they were dead because the
position index was never wired. `iterate` re-read the WHOLE directory on
every call and discarded every entry below the caller's position, so a
full walk of a directory cost time quadratic in its entry count. The
index that avoids it existed and had no callers.

`read_dir_from` seeds the walk from that index, and `iterate` uses it.
Wiring it exposed a latent off-by-three: the index returned the caller's
position as the resume length whenever that position was at or below the
synthetic entries, mixing an external position into a variable the walk
spends as an on-disk byte count. That pushes every entry's reported
position three too high on an ordinary full listing. The reference keeps
one convention throughout and adds the three at the end; this port
subtracts it, so the early return is zero.

+++ crates/kernel/squashfs/src/mount/sb.rs
@@ -34,7 +34,7 @@ impl SuperOps for SquashSuperOps {
             // none to report; inventing one would collide with a real fsid.
             f_fsid: 0,
             f_flags: 0,
-            f_namelen: NAME_LEN as u32,
+            f_namelen: NAME_LEN as u64,
             f_frsize: 0,
         })
     }
